Refactor signup flow with onboarding layout and terms & conditions page
