+++ .env.example
@@ -4,6 +4,7 @@ NEXT_PUBLIC_METADATA_ASSETS_S3_BUCKET=
 NEXT_PUBLIC_CLOUDFLARE_RESIZE_PREFIX=
 NEXT_PUBLIC_ONLYDUST_API_BASEPATH=
 NEXT_PUBLIC_MARKETPLACE_URL=
+NEXT_PUBLIC_LEGALS_S3_BUCKET=
 
 ######################################## AUTH0 ########################################
